+++ package.json
@@ -24,6 +24,7 @@
         "@types/node": "^18.15.5",
         "@types/yargs": "^17.0.23",
         "@types/split2": "^3.2.1",
+        "@types/ws": "^8.5.4",
         "eslint": "^8.36.0",
         "eslint-config-prettier": "^8.8.0",
         "eslint-plugin-import": "^2.26.0",
