+++ src/config.ts
@@ -1,3 +1,9 @@
+import * as fs from "fs";
+import * as JSON5 from "json5";
+import * as yargs from "yargs";
+import * as ConfigSchema from "../schema/Config.schema.json";
+import { Validator } from "jsonschema";
+
 /** Bot config */
 export interface Config {
     /**
@@ -8,12 +14,19 @@ export interface Config {
     url?: string;
     /** Bot username */
     username: string;
+    /** Bot ID. This is automatically set.
+     * @hidden
+     */
+    bot_id?: number;
     /** API key for the bot. */
     apikey: string;
     /** Config for how to run your bot */
     bot?: BotConfig;
     opening_bot?: BotConfig;
     resign_bot?: BotConfig;
+
+    /** Send a message saying what the bot thought the score was at the end of the game */
+    farewellscore?: boolean;
 
     log_file?: string;
 
@@ -58,7 +71,7 @@ export interface BotConfig {
     command: string;
 }
 
-export function defaults(): Config {
+function defaults(): Config {
     return {
         username: "",
         apikey: "",
@@ -67,8 +80,35 @@ export function defaults(): Config {
     };
 }
 
-export const config: Config = defaults();
+export const config: Config = load_config_or_exit();
 
-export function set_config(new_config: Config) {
-    Object.assign(config, new_config);
+function load_config_or_exit(): Config {
+    yargs(process.argv.slice(2))
+        .describe("url", "URL of the OGS server to connect to, defaults to https://online-go.com")
+        .parseSync();
+
+    const filename = "test.json5";
+
+    /* eslint-disable-next-line @typescript-eslint/no-var-requires */
+    const contents = fs.readFileSync(filename, "utf8");
+    const raw = JSON5.parse(contents);
+    const with_defaults = { ...defaults(), ...raw };
+    const validator = new Validator();
+    const result = validator.validate(with_defaults, ConfigSchema);
+
+    if (!result.valid) {
+        console.error(``);
+        console.error(``);
+        console.error(`Invalid config file: ${filename}`);
+
+        for (const error of result.errors) {
+            console.error(`\t ${error.toString()}`);
+        }
+        console.error(``);
+        console.error(``);
+
+        process.exit(1);
+    }
+
+    return with_defaults;
 }
